+++ designer/client/src/components/graph/node-modal/editors/expression/Table/state/action.ts
@@ -12,6 +12,7 @@ export const ActionTypes = {
     columnResize: "column-resize",
     renameColumn: "rename-column",
     changeColumnType: "change-column-type",
+    moveColumn: "move-column",
 } as const;
 
 type Actions = {
@@ -57,6 +58,10 @@ type Actions = {
     [ActionTypes.changeColumnType]: {
         column: number;
         dataType: string;
+    };
+    [ActionTypes.moveColumn]: {
+        startIndex: number;
+        endIndex: number;
     };
 };