+++ security/src/main/scala/pl/touk/nussknacker/ui/security/basicauth/BasicAuthenticationResources.scala
@@ -14,8 +14,7 @@ class BasicAuthenticationResources(
     override val configuration: BasicAuthenticationConfiguration
 )(
     implicit executionContext: ExecutionContext
-) extends AuthenticationResources
-    with AnonymousAccess {
+) extends AuthenticationResources {
 
   override type CONFIG = BasicAuthenticationConfiguration
 
@@ -23,18 +22,34 @@ class BasicAuthenticationResources(
 
   override protected val frontendStrategySettings: FrontendStrategySettings = FrontendStrategySettings.Browser
 
-  override protected def authenticateReally(): AuthenticationDirective[AuthenticatedUser] =
+  override def authenticate(): AuthenticationDirective[AuthenticatedUser] =
     SecurityDirectives.authenticateBasicAsync(
       authenticator = authenticator,
       realm = realm
     )
 
-  override protected def authenticateReally(credentials: PassedAuthCredentials): Future[Option[AuthenticatedUser]] = {
-    authenticator.authenticate(credentials)
+  override def authenticate(authCredentials: PassedAuthCredentials): Future[Option[AuthenticatedUser]] = {
+    authenticator.authenticate(authCredentials)
   }
 
-  override protected def rawAuthCredentialsMethod: EndpointInput[Option[String]] = {
-    auth.basic[Option[String]](WWWAuthenticateChallenge.basic.realm(realm))
+  override def authenticationMethod(): EndpointInput[Option[PassedAuthCredentials]] =
+    auth
+      .basic[Option[String]](WWWAuthenticateChallenge.basic.realm(realm))
+      .map(_.map(PassedAuthCredentials))(_.map(_.value))
+
+  override def impersonationSupport: ImpersonationSupport = new ImpersonationSupported {
+
+    override def getImpersonatedUserData(impersonatedUserIdentity: String): Option[ImpersonatedUserData] =
+      configuration.users
+        .find { _.identity == impersonatedUserIdentity }
+        .flatMap { configUser =>
+          configUser.username match {
+            case Some(username) => Some(ImpersonatedUserData(configUser.identity, username, configUser.roles))
+            case None => Some(ImpersonatedUserData(configUser.identity, configUser.identity, configUser.roles))
+          }
+        }
+
   }
 
+  override def getAnonymousRole: Option[String] = configuration.anonymousUserRole
 }